Add doc of modal
